Implement workaround should patching \author fail
In some combinations of LaTeX distribution and document class, the
macros \author, \title, and \date seem to have definitions that contain
two spaces. As a consequence, they can not be cleanly retokenized and
thus not patched. Should this be the case we instead provide alternative
commands like \frontpageauthor to set the author for the frontpage.